refactor: move util to other file
